separate test cases in trees and other stuff
